kata-solutions/questionnaire-kata/questionnaire-python/pyproject.toml
[tool.poetry]
name = "DataProcessor"
version = "0.1.0"
description = "This kata practices TDD"
authors = ["Your name  <your@email.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"

pytest-watch = "^4.2.0"
pytest = "^7.3.2"
pyhamcrest = "^2.0.4"
pandas = "^2.1.1"
openpyxl = "^3.1.2"
streamlit = "^1.27.2"
plotly-express = "^0.4.1"




[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
